Now works for all 16 / 9 Resolutions :)

+++ utils.py
@@ -105,10 +105,11 @@ def nice_name(name):
 
 
 # Transforms from 1920x1080 Base
+# Works for all 16 / 9 Resolutions
 def transform_coordinates(handle, x, y):
     x1, y1, x2, y2 = win32gui.GetWindowRect(handle)
     w = x2 - x1
     h = y2 - y1
-    new_x = (w / 1920) * x
-    new_y = (h / 1080) * y
-    return new_x, new_y
+    new_x = int((w / 1920) * x)
+    new_y = int((h / 1080) * y)
+    return (new_x, new_y)
